Revert "scanner: store result on callback error"
This reverts commit cf3e869aa40311feb32597bcced2e8edcc9ab6c1.

The errors raised in callback_error are programming errors since
any exception should have created already a Result or ignored.
It will be better to realize whether new exceptions are raised
while programming that to discover them time later in the bandwidth
files.
The possible errors at the moment have tickets and there should
not be created new ones.
Callback_error should be also removed to raise a possible exception
with the full trace.