fix: relaylist: Check exit to all domains/ips
When an exit policy allows to exit only to some subnet, it is not
enough to check that it can exit to a port, since it can, but it might
not be able to exit to the domain/ip of the sbws Web servers.
To ensure that without having to check whether it can exit to a
specific domain/ip, we can query the exit policy with `strict`.

Closes #400006. Bugfix v1.0.3.